Extract AlternateConnectionExecutor as separate class.
