+++ liquibase-integration-tests/src/test/java/liquibase/dbtest/h2/H2IntegrationTest.java
@@ -28,6 +28,7 @@ import liquibase.executor.ExecutorService;
 import liquibase.executor.jvm.JdbcExecutor;
 import liquibase.extension.testing.testsystem.DatabaseTestSystem;
 import liquibase.extension.testing.testsystem.TestSystemFactory;
+import liquibase.helper.AlternateConnectionExecutor;
 import liquibase.precondition.core.RowCountPrecondition;
 import liquibase.precondition.core.TableExistsPrecondition;
 import liquibase.sql.visitor.SqlVisitor;
@@ -268,41 +269,6 @@ public class H2IntegrationTest extends AbstractIntegrationTest {
         assertEquals(1, ex.getFailedPreconditions().size());
         String expectedMessage = String.format("%s does not exist", tableName);
         assertTrue(ex.getFailedPreconditions().get(0).getMessage().endsWith(expectedMessage));
-    }
-
-    /**
-     * An {@link JdbcExecutor} that provides its own {@link DatabaseConnection} and commits after executing
-     * each change
-     */
-    public class AlternateConnectionExecutor extends JdbcExecutor {
-
-        public AlternateConnectionExecutor() throws Exception {
-            String urlParameters = ";INIT=CREATE SCHEMA IF NOT EXISTS lbschem2\\;SET SCHEMA lbschem2";
-            DatabaseTestSystem testSystem = (DatabaseTestSystem) Scope.getCurrentScope().getSingleton(TestSystemFactory.class).getTestSystem("h2");
-            database = DatabaseFactory.getInstance().openDatabase(testSystem.getConnectionUrl().replace("lbcat", "lbcat2") + urlParameters,
-                    testSystem.getUsername(), testSystem.getPassword(), null, null);
-            database.setAutoCommit(false);
-        }
-
-        @Override
-        public String getName() {
-            return "h2alt";
-        }
-
-        @Override
-        public void setDatabase(Database database) {
-            // ignore the database connection passed in since we're providing our own
-        }
-
-        public Database getDatabase() {
-            return database;
-        }
-
-        @Override
-        public void execute(Change change, List<SqlVisitor> sqlVisitors) throws DatabaseException {
-            super.execute(change, sqlVisitors);
-            database.commit();
-        }
     }
 
     @Override
